+++ connector/src/test/java/it/eng/connector/integration/datatransfer/DataTransferAPIDownloadDataIT.java
@@ -215,6 +215,7 @@ public class DataTransferAPIDownloadDataIT extends BaseIntegrationTest {
         assertEquals(transferProcessStarted.getAgreementId(), transferProcessFromDb.getAgreementId());
         assertEquals(transferProcessStarted.getCallbackAddress(), transferProcessFromDb.getCallbackAddress());
         assertEquals(TransferState.COMPLETED, transferProcessFromDb.getState());
+        assertFalse(transferProcessFromDb.isDownloading());
         // +1 from test
         assertEquals(startingTransferProcessCollectionSize + 1, transferProcessRepository.findAll().size());
     }
@@ -336,6 +337,7 @@ public class DataTransferAPIDownloadDataIT extends BaseIntegrationTest {
         assertEquals(transferProcessStarted.getAgreementId(), transferProcessFromDb.getAgreementId());
         assertEquals(transferProcessStarted.getCallbackAddress(), transferProcessFromDb.getCallbackAddress());
         assertEquals(TransferState.COMPLETED, transferProcessFromDb.getState());
+        assertFalse(transferProcessFromDb.isDownloading());
         // +1 from test
         assertEquals(startingTransferProcessCollectionSize + 1, transferProcessRepository.findAll().size());
     }
@@ -426,6 +428,7 @@ public class DataTransferAPIDownloadDataIT extends BaseIntegrationTest {
         TransferProcess transferProcessFromDb = transferProcessRepository.findById(transferProcessStarted.getId()).get();
 
         assertFalse(transferProcessFromDb.isDownloaded());
+        assertFalse(transferProcessFromDb.isDownloading());
         assertNull(transferProcessFromDb.getDataId());
         assertEquals(transferProcessStarted.getConsumerPid(), transferProcessFromDb.getConsumerPid());
         assertEquals(transferProcessStarted.getProviderPid(), transferProcessFromDb.getProviderPid());